# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BAPS Anand Jamo Ne Jamadu Seva Portal API
  description: Module 3 (Books & Receipts) OpenAPI specification
  version: 3.0.0
  contact:
    name: BAPS Anand Mandir
servers:
  - url: http://localhost:5000/api/v1
    description: Local development server
security:
  - CookieAuth: []
  - BearerAuth: []
paths:
...
  /auth/users:
    get:
      summary: List all users (SUPER_ADMIN only)
      operationId: listUsers
      tags:
        - Auth
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create a new user (SUPER_ADMIN only)
      operationId: createUser
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Validation error or invalid role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /prasad/distribute:
    post:
      summary: Record prasad distribution for a sevak
      operationId: distributePrasad
      tags:
        - Prasad
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrasadDistributeRequest'
      responses:
        '200':
          description: Distribution recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrasadDistributeResponse'
        '400':
          description: Validation or over-allocation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Receipts already served
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /reports/master-summary/excel:
    get:
      summary: Download master summary Excel workbook
      operationId: exportMasterSummaryExcel
      tags:
        - Reports
      responses:
        '200':
          description: Excel file download
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateUserRequest:
      type: object
      required:
        - email
        - password
        - fullName
        - role
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        fullName:
          type: string
        phoneNumber:
          type: string
          nullable: true
        role:
          type: string
          enum:
            - SUPER_ADMIN
            - VOLUNTEER
    User:
      type: object
      required:
        - id
        - email
        - fullName
        - role
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        fullName:
          type: string
        phoneNumber:
          type: string
          nullable: true
        role:
          type: string
          enum:
            - SUPER_ADMIN
            - VOLUNTEER
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserResponse:
      type: object
      required:
        - user
      properties:
        user:
          $ref: '#/components/schemas/User'
...
    UserListResponse:
      type: object
      required:
        - users
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
...
    PrasadDistributeRequest:
      type: object
      required:
        - sevakId
        - tierDistributions
        - receiptIdsToMark
      properties:
        sevakId:
          type: string
          format: uuid
        tierDistributions:
          type: array
          items:
            type: object
            required:
              - slabTier
              - distributeCount
            properties:
              slabTier:
                type: string
              distributeCount:
                type: integer
                minimum: 0
        receiptIdsToMark:
          type: array
          items:
            type: string
            format: uuid
    PrasadDistributeResponse:
      type: object
      required:
        - message
        - receiptsServed
      properties:
        message:
          type: string
        receiptsServed:
          type: integer
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string